split joint + rigid_body
